BUG FIX: FIX may not be sourced and the wrong FIX labels are referenced
Created helper function to source FIX directory prior to system call. This
was not guarenteed to be performed prior to the system call.
Additionally, the incorrect FIX labels file was referenced for the FIX apply
step. This would cause FIX to look for the necessary files in that same directory,
but not finding them. This was remedied by copyting the label FIX file back to the
FIX processing directory.